Checkins in progress now simply store the topicId in the Project object. Since Project objects belong to the "Store" schema, they are stored transactionally, so the BDB transaction system takes care of concurrent access.
On BIMserver startup this field (Project.chekinInProgress) is now used
to determine which projects should be cleaned-up, this will be much
faster than before since only projects that had checkins going on need
to be checked. Previously this was stored in memory. One advantage of
this method is that when bugs in BIMserver disallow checkin-in on a
project, restarting the server can (temporarely) solve the problem
(until the bug is fixed).

https://github.com/opensourceBIM/BIMserver/issues/755

+++ BimServer/src/org/bimserver/cache/NewDiskCacheWriter.java
@@ -57,7 +57,9 @@ public class NewDiskCacheWriter implements Writer, DiskCacheItem, Closeable {
 		LOGGER.info("Renaming temp file " + tempFile.getFileName().toString() + " to " + file.getFileName().toString());
 		outputStream.writeInt(-1);
 		outputStream.close();
-		Files.move(tempFile, file);
+		if (Files.exists(tempFile)) {
+			Files.move(tempFile, file);
+		}
 		newDiskCacheManager.doneGenerating(this);
 		latch.countDown();
 	}
